split single header file into .h & .c
