+++ src/csp.h
@@ -4,7 +4,6 @@
 #ifndef CSP_CSP_H
 #define CSP_CSP_H
 
-#include <stdlib.h>
 #include "htslib/sam.h"
 
 /*
@@ -24,16 +23,9 @@ typedef struct {
 @note    The pointer returned successfully by csp_bam_fs_init() should be freed
          by csp_bam_fs_destroy() when no longer used.
  */
-static inline csp_bam_fs* csp_bam_fs_init(void) { return (csp_bam_fs*) calloc(1, sizeof(csp_bam_fs)); }
+static inline csp_bam_fs* csp_bam_fs_init(void);
 
-static inline void csp_bam_fs_destroy(csp_bam_fs *p) {
-    if (p) {
-        if (p->idx) { hts_idx_destroy(p->idx); }
-        if (p->hdr) { sam_hdr_destroy(p->hdr); }
-        if (p->fp)  { hts_close(p->fp); }
-        free(p);
-    }
-}
+static inline void csp_bam_fs_destroy(csp_bam_fs *p);
 
 /*@abstract  Build the csp_bam_fs structure.  
 @param fn    Filename of the bam/sam/cram.
@@ -44,29 +36,8 @@ static inline void csp_bam_fs_destroy(csp_bam_fs *p) {
 @note        The pointer returned successfully by csp_bam_fs_build() should be freed
              by csp_bam_fs_destroy() when no longer used.
 */
-static inline csp_bam_fs* csp_bam_fs_build(const char *fn, int *ret) {
-    csp_bam_fs *p;
-    if (NULL == fn) { *ret = -1; return NULL; }
-    if (NULL == (p = csp_bam_fs_init())) { *ret = -1; return NULL; }
-    if (NULL == (p->fp = hts_open(fn, "rb"))) { *ret = -2; goto fail; }
-    if (NULL == (p->hdr = sam_hdr_read(p->fp))) { *ret = -2; goto fail; }
-    if (NULL == (p->idx = sam_index_load(p->fp, fn))) { *ret = -2; goto fail; }
-    *ret = 0;
-    return p;
-  fail:
-    csp_bam_fs_destroy(p);
-    return NULL;		
-}
+static inline csp_bam_fs* csp_bam_fs_build(const char *fn, int *ret);
 
-static inline int csp_bam_fs_reset(csp_bam_fs *p, const char *fn) {
-    if (NULL == p) { return -1; }
-    if (p->idx) { hts_idx_destroy(p->idx); }
-    if (p->hdr) { sam_hdr_destroy(p->hdr); }
-    if (p->fp)  { hts_close(p->fp); }		
-    if (NULL == (p->fp = hts_open(fn, "rb"))) { return -1; }
-    if (NULL == (p->hdr = sam_hdr_read(p->fp))) { return -1; }
-    if (NULL == (p->idx = sam_index_load(p->fp, fn))) { return -1; }
-    return 0;
-}
+static inline int csp_bam_fs_reset(csp_bam_fs *p, const char *fn);
 
 #endif
