+++ src/jstring.h
@@ -5,7 +5,6 @@
 #define SZ_JSTRING_H
 
 #include <stdlib.h>
-#include <string.h>
 #include "htslib/kstring.h"        // do not use "kstring.h" as it's different from "htslib/kstring.h"
 
 /*
@@ -23,11 +22,7 @@
 @param n       Num of char* elements in the array.
 @return        Void.
  */
-static inline void str_arr_destroy(char **a, const int n) {
-    int i;      
-    for (i = 0; i < n; i++) free(a[i]);
-    free(a);
-}
+static inline void str_arr_destroy(char **a, const int n);
 
 /*abstract    Join an array of strings (char*) by a delim char.
 @param a      Pointer to the char* array.
@@ -36,11 +31,6 @@ static inline void str_arr_destroy(char **a, const int n) {
 @param s      Pointer of the kstring_t structure to save the result.
 @return       Num of elements that are successfully joined.
  */
-static inline int str_arr_join(char **a, const int n, int c, kstring_t *s) {
-    int i;      /* TODO: to test if the ret of ksxxx functions is right. */	
-    for (i = 0; i < n - 1; i++) { kputs(a[i], s); kputc_(c, s); }
-    if (n >= 1) { kputs(a[i], s); }
-    return n;
-}
+static inline int str_arr_join(char **a, const int n, int c, kstring_t *s);
 
 #endif
